change to catch not modified response and get config working correctly
