+++ src/Ocelot/DependencyInjection/ServiceCollectionExtensions.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using CacheManager.Core;
-using IdentityServer4.AccessTokenValidation;
 using IdentityServer4.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
@@ -61,7 +60,7 @@ namespace Ocelot.DependencyInjection
             services.AddSingleton<IConfigurationValidator, FileConfigurationValidator>();
             services.AddSingleton<IBaseUrlFinder, BaseUrlFinder>();
 
-            var identityServerConfiguration = GetIdentityServerConfiguration();
+            var identityServerConfiguration = IdentityServerConfigurationCreator.GetIdentityServerConfiguration();
             
             if(identityServerConfiguration != null)
             {
@@ -141,30 +140,6 @@ namespace Ocelot.DependencyInjection
             services.AddScoped<IRequestScopedDataRepository, HttpDataRepository>();
 
             return services;
-        }
-
-        private static IdentityServerConfiguration GetIdentityServerConfiguration()
-        {
-            var username = Environment.GetEnvironmentVariable("OCELOT_USERNAME");
-            var hash = Environment.GetEnvironmentVariable("OCELOT_HASH");
-            var salt = Environment.GetEnvironmentVariable("OCELOT_SALT");
-
-            return new IdentityServerConfiguration(
-                "admin",
-                false,
-                SupportedTokens.Both,
-                "secret",
-                new List<string> {"admin", "openid", "offline_access"},
-                "Ocelot Administration",
-                true,
-                GrantTypes.ResourceOwnerPassword,
-                AccessTokenType.Jwt,
-                false,
-                new List<User> 
-                {
-                    new User("admin", username, hash, salt)
-                }
-            );
         }
     }
 }
